Clean up Data Source attributes.
